feat: add initial project structure with models, routes, controllers, and configuration
