+++ src/config/db.js
@@ -1,8 +1,9 @@
 import mysql from "mysql2";
 import knex from "knex";
-import { knexconfig } from "./knexconfig.js";
-
 import dotenv from "dotenv";
+import { knexconfig } from '../config/knexConfig.js';
+
+
 dotenv.config();
 
 const DB_HOST = process.env.DB_HOST;
