+++ src/models/authModel.js
@@ -1,3 +1,5 @@
+// import { db } from "../config/db.js";
+
 import { db } from "../config/db.js";
 
 const userAccountTable = "hris_user_accounts";
