+++ src/routes/authRoutes.js
@@ -1,0 +1,27 @@
+import express from "express";
+import {
+  getServices,
+  login,
+  logout,
+  refreshToken,
+  userInfo,
+} from "../controllers/authController.js";
+import verifyJWT from "../middlewares/verifyJWT.js";
+
+const router = express.Router();
+
+router.post("/login", login);
+
+// router.post("/logout", logout);
+
+// router.get("/user-info", verifyJWT, userInfo);
+
+// router.get("/get-services/:id", getServices);
+
+// router.get("/refresh-token", refreshToken);
+
+// router.get("/profile", verifyJWT, (req, res) => {
+//   return res.json({ message: "Profile data", user: req.user });
+// });
+
+export default router;
